+++ src/interpreter/value.rs
@@ -71,7 +71,7 @@ impl Display for Value {
             Value::Addr(a) => write!(f, "{}", a.to_checksum(None)),
             Value::Str(s) => write!(f, "\"{}\"", s),
             Value::FixBytes(w, s) => {
-                let bytes = w[..*s].to_vec();
+                let bytes = w[32 - *s..].to_vec();
                 write!(f, "0x{}", hex::encode(bytes))
             }
             Value::Bytes(bytes) => write!(f, "0x{}", hex::encode(bytes)),
@@ -292,8 +292,9 @@ impl FromHex for Value {
         } else if hex.as_ref().len() == 42 {
             Value::Addr(Address::from_hex(hex)?)
         } else if hex.as_ref().len() <= 66 {
-            let mut bytes = Vec::from_hex(&hex.as_ref()[2..])?;
-            bytes.resize(32, 0);
+            let data = Vec::from_hex(&hex.as_ref()[2..])?;
+            let mut bytes = vec![0; 32];
+            bytes[32 - data.len()..].copy_from_slice(&data);
             Value::FixBytes(B256::from_slice(&bytes), (hex.as_ref().len() - 2) / 2)
         } else {
             Value::Bytes(Vec::from_hex(&hex.as_ref()[2..])?)
@@ -748,8 +749,9 @@ mod tests {
         assert_eq!(value, Value::Addr(addr));
 
         let value = Value::from_hex("0xdeadbeef").unwrap();
-        let mut bytes = Vec::from_hex("deadbeef").unwrap();
-        bytes.resize(32, 0);
+        let bytes =
+            Vec::from_hex("00000000000000000000000000000000000000000000000000000000deadbeef")
+                .unwrap();
         let fix_bytes = B256::from_slice(&bytes);
         assert_eq!(value, Value::FixBytes(fix_bytes, 4));
     }