+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Not released
+## v0.1.4 (2025-01-26)
 
 ### Features